+++ app/lib/prism.ts
@@ -2,6 +2,10 @@ import Prism from "prismjs";
 
 import "prismjs/components/prism-powershell";
 import "prismjs/components/prism-json";
+import "prismjs/components/prism-javascript";
+import "prismjs/components/prism-typescript";
+import "prismjs/components/prism-jsx";
+import "prismjs/components/prism-tsx";
 import "./prism.css";
 
 export function highlightCode(code: string, language: string) {